pyproject.toml
[project]
name = "LLMQuizChecker"
version = "0.1.0"
description = "Extracts quiz answers (MCQ, MSQ, True/False) from scanned answer sheets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easygui>=0.98.3",
    "google-generativeai>=0.8.3",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]
